Attempted to get evalPoly to return numbers as elements of QQ
evalPoly will return numbers as numbers, but this breaks 'support' command, so I added a way for support to ignore elements in QQ in the list (i'd rather have it return {} for those but this works). I thought this would help but it doesn't seem to have made a difference to other errors. also specified "nonempty" for samplePoints and added comments on lazardProjection